Move abbrev related tests to AbbreviationMatcherTest
